clientServerTUI
This is a big change, with many new files. That's because I didn't just
implement a new standalone pure framework, I've tried to write some
reusable pieces and then to plumb them together to create either
playTUI, clientServerTUI, or a future client-server pure framework which
uses something other than a TUI. Those reusable pieces turned out to
have different shapes, which I named "threads", "stream framework", and
"io framework". I'm not 100% happy with the result, but it's a start.

One thing I am especially not happy with is the clientServerTUI type. To
match the playTUI type, it takes a bunch of pure functions which return
a new world, and since we need to be able to send messages over the
network in response to an event, it now also returns a list of messages.
The functions I wrote which return a list of messages and an updated
world just beg to be rewritten to (WritetT msg State world). But if I
change the type of clientServerTUI to take monadic computations instead
of pure functions, should I also change playTUI to match, even though it
will no longer match the type of gloss's play?