multiplayTUI
building upon clientServerTUI to provide the nice interface presented in
the blog post.

this is the right interface, and it works, but the user experience is
still sub-optimal if the latency is poor, as the client's screen does
not react immediately after the user types something, only once the
server confirms the keypress.