react immediately to the user's keypresses
even if the actual result might change slightly once the server's
authoritative sequence of actions comes in. for example, in a game in
which the first player to type a key wins and both players press it at
around the same time, both players will immediately see a win screen,
then the keys are sent to the server, the server receives them in some
order, it sends the keys back in the same order, and now the screen of
both players update to reflect the correct winner.